requirements.txt
opencv-python
psutil
python-dotenv
groq
AppOpener
pywhatkit
bs4
rich
keyboard
cohere
googlesearch-python
selenium
mtranslate
pygame
edge-tts
PyQt5
webdriver-manager
ddgs
tzlocal
pycaw
comtypes
screen-brightness-control
aiofiles
github